Antennas

Documentation for Antennas.

Antennas.AntennaArray

Antennas.AntennaArray

Antennas.RadiationPattern

Antennas.RadiationPattern

Antennas.ArrayFactor

Antennas.baselines

Antennas.calibrate

Antennas.read_HFSS_pattern

Antennas.AntennaArray — Type

AntennaArray

Holds the antenna array data, isa AbstractDimArray

Antennas.AntennaArray — Method

aa(ϕ,θ,freq)

Finds the relative amplitude of the array factor for the array aa in the azimuth ϕ, elevation θ at frequency freq. Angles in degrees.

Antennas.RadiationPattern — Type

RadiationPattern

Holds the pattern data, isa AbstractDimArray

Antennas.RadiationPattern — Type

RadiationPattern(f::Function, dim::Dimension [, name])

Apply function f across the values of the dimension dim (using broadcast), and return the result as a RadiationPattern with the given dimension. Optionally provide a name for the result.

Antennas.ArrayFactor — Method

ArrayFactor(aa)

Constructs a RadiationPattern representative of the array factor of the array aa sampled over azimuth ϕ and elevation θ at frequenc[y|ies] freq.

Antennas.baselines — Method

" baselines(af) Returns all the baseline vectors for the antenna locations in aa.

Antennas.calibrate — Method

calibrate(array,phase_center)

For a given antenna array array, return a new AntennaArray whose phase center has been adjusted to phase_center.

Antennas.read_HFSS_pattern — Method

read_HFSS_pattern("myAntenna.csv")

Utility function to read the exported fields from HFSS into a RadiationPattern.